reorg regtests based on source file to be tested and map directory structure similar to source code
