--- tests/test_inspect.py
@@ -1,10 +1,0 @@
-from buildtest.menu.config import show_configuration
-from buildtest.menu.show import show_schema_layout
-
-
-def test_show_config():
-    show_configuration()
-
-
-def test_show_schema():
-    show_schema_layout()
