+++ tests/utils/test_command.py
